--- tests/subDocuments/data/comments.json
@@ -1,9 +1,0 @@
-{
-	"one": {
-		"name": "First Comment"
-	},
-	"two": {
-  "name": "Second Comment",
-  "for": { "article": "First Post", "onModel": "Post" }
-	}
-}
