--- tests/subDocuments/data/posts.json
@@ -1,8 +1,0 @@
-{
-	"one": {
-		"name": "First Post"
-	},
-	"two": {
-		"name": "Second Post"
-	}
-}
