--- tests/subDocuments/models/Post.js
@@ -1,4 +1,0 @@
-const mongoose = require("mongoose");
-const { Schema } = mongoose;
-
-module.exports = mongoose.model("Post", new Schema({ name: String }));
